+++ infretis/classes/engines/factory.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Any
 from infretis.classes.engines.cp2k import CP2KEngine
 from infretis.classes.engines.gromacs import GromacsEngine
 from infretis.classes.engines.lammps import LAMMPSEngine
-from infretis.classes.engines.turtlemd import TurtleMDEngine
+from infretis.classes.engines.turtlemdengine import TurtleMDEngine
 from infretis.core.core import create_external, generic_factory
 
 if TYPE_CHECKING:  # pragma: no cover
